requirements.txt
boto3
nfllivepy==0.1.0a1
python-dotenv
tweepy